test: test stac extension from package
