+++ src/stactools/hotosm/stac.py
@@ -36,6 +36,11 @@ from stactools.hotosm.oam_metadata import OamMetadata
 ALTERNATE_ASSETS_VERSION = "v1.2.0"
 ALTERNATE_ASSETS_SCHEMA = f"https://stac-extensions.github.io/alternate-assets/{ALTERNATE_ASSETS_VERSION}/schema.json"
 
+OAM_EXT_VERSION = "v0.1.0"
+OAM_EXT_SCHEMA = (
+    f"https://hotosm.github.io/stactools-hotosm/oam/{OAM_EXT_VERSION}/schema.json"
+)
+
 
 def create_collection() -> Collection:
     """Create a STAC Collection of OAM imagery."""
@@ -210,6 +215,7 @@ def create_item(oam_metadata: OamMetadata) -> Item:
     )
 
     item.validate()
+
     return item
