+++ tests/conftest.py
@@ -4,9 +4,12 @@ import datetime as dt
 import json
 from pathlib import Path
 
+import numpy as np
 import pyproj
 import pytest
+import rasterio
 import shapely
+from rasterio.transform import from_bounds
 from shapely.wkt import dumps as wkt_dumps
 
 from stactools.hotosm.oam_metadata import OamMetadata
@@ -59,3 +62,33 @@ def example_oam_metadata() -> OamMetadata:
 def example_oam_meta_api_response() -> dict:
     """Example OAM Metadata API response."""
     return json.loads((DATA / "oam_meta_api_response10.json").read_text())
+
+
+@pytest.fixture
+def example_oam_image(example_oam_metadata: OamMetadata, tmp_path) -> OamMetadata:
+    """Provision data for the example OAM metadata item."""
+    data = np.random.randint(low=0, high=255, size=(3, 13, 42), dtype="uint8")
+    profile = {
+        "count": data.shape[0],
+        "height": data.shape[1],
+        "width": data.shape[2],
+        "dtype": "uint8",
+        "crs": rasterio.crs.CRS.from_string(example_oam_metadata.projection_wkt),
+        "transform": from_bounds(*example_oam_metadata.bbox, *data.shape[1:]),
+    }
+
+    example_oam_metadata.image_url = str(tmp_path / example_oam_metadata.image_url)
+    with rasterio.open(
+        example_oam_metadata.image_url, "w", driver="COG", **profile
+    ) as dst:
+        dst.write(data)
+
+    example_oam_metadata.thumbnail_url = str(
+        tmp_path / example_oam_metadata.thumbnail_url
+    )
+    with rasterio.open(
+        example_oam_metadata.thumbnail_url, "w", driver="PNG", **profile
+    ) as dst:
+        dst.write(data)
+
+    return example_oam_metadata
